Add another 'undefined_variable' test
This test was built when I was looking into possibility of
using register variable for storing instruction pointer. Apart
from manually updating frames, we also need to supply it
in runTimError calls so that was too much plumbing so I decided
to abandon this idea. It adds quite a bit of complexity.